requirements.txt
PyYAML>=6.0
requests>=2.31.0
faster-whisper>=0.10.0
google-api-python-client>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
schedule>=1.2.0
